requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
pytest>=7.4.0
scipy>=1.10.0
scikit-learn>=1.2.0
tensorflow>=2.12.0
pytorch>=2.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
aiohttp>=3.8.4
websockets>=11.0.3 